spec copier
https://coveord.atlassian.net/browse/CDX-1061
