+++ packages/cli-e2e/jest.config.ts
@@ -3,11 +3,13 @@
  * https://jestjs.io/docs/en/configuration.html
  */
 
+const testOsFilter = process.platform === 'win32' ? 'windows' : 'linux';
+
 export default {
   testEnvironment: 'node',
   globalSetup: process.env.CI ? './setup/ci.ts' : './setup/local.ts',
   globalTeardown: './teardown.ts',
   testMatch: process.env.CI
-    ? ['**/__tests__/**/*specs?(.ci).ts']
-    : ['**/__tests__/**/*specs.ts'],
+    ? [`**/__tests__/**/*specs?(.ci|.${testOsFilter}).ts`]
+    : [`**/__tests__/**/*specs?(.${testOsFilter}).ts`],
 };
